refactor s2anet to support batch_size > 1
