+++ configs/dota/s2anet_1x_spine.yml
@@ -13,17 +13,6 @@ LearningRate:
   base_lr: 0.01
 
 S2ANetHead:
-  anchor_strides: [8, 16, 32, 64, 128]
-  anchor_scales: [4]
-  anchor_ratios: [1.0]
-  anchor_assign: RBoxAssigner
-  stacked_convs: 2
-  feat_in: 256
-  feat_out: 256
-  num_classes: 9
-  align_conv_type: 'AlignConv'  # AlignConv Conv
-  align_conv_size: 3
-  use_sigmoid_cls: True
   reg_loss_weight: [1.0, 1.0, 1.0, 1.0, 1.05]
   cls_loss_weight: [1.05, 1.0]
-  reg_loss_type: 'l1'
+
